refactor(embedding): extract composition-root/selection seam (#173)
Fowler Move Function / Extract Class. Introduce embedding_factory.py — the
"engine selection / wiring" seam (brief seam 3): the single place that reads
the runtime settings (EMBEDDING_DIM / EMBEDDING_DEVICE) and assembles the
process-wide encoder. get_embedding_engine / reset_embedding_engine move here;
embedding_engine re-exports them so every existing consumer import
(`from ...embedding_engine import get_embedding_engine`) keeps working.

The concrete EmbeddingEngine + get_memory_settings are imported lazily inside
get_embedding_engine so the factory has no module-load dependency on the engine
(no import cycle; verified importing in either order).

Result — the three seams are now three modules plus the concrete provider:
  embedding_provider.py   115  (seam 1: interface)
  embedding_model_lifecycle.py 294  (seam 2: load state machine)
  embedding_factory.py     52  (seam 3: selection/composition root)
  embedding_engine.py     245  (concrete neural provider)
All <=300 lines, every method <=40. Behaviour-preserving; embedding tests pass
unchanged.
